AI-Powered Search Engine/requirements.txt
sentence-transformers==2.5.1
faiss-cpu==1.8.0
pandas==2.2.1
numpy==1.26.4
flask==3.0.2
tqdm==4.66.2
scikit-learn==1.4.2